+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 # CHANGELOG
 
+## Version 2.6.1
+* Added support for token-based authentication as an alternative to username and password (ORION-226852)
+
 ## Version 2.6.0
 * Block CSI Driver (VCSI-193)
+
+## Version 2.5.1
+* Custom driver name
+* Allow optional qos_policy_id storage class argument (unsupported via helm) (VCSI-226)
 
 ## Version 2.5.0
 * CSI driver operator (VCSI-173)